% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{astro}
\alias{astro}
\title{astro}
\usage{
astro(t_astro, simon, longitude, hours, ddt)
}
\arguments{
\item{t_astro}{astronomical time}

\item{simon}{coefficient matrix}

\item{longitude}{the longitude}

\item{hours}{hour or measurement}

\item{ddt}{time in ddt}
}
\value{
astonomical parameters
}
\description{
Calculate astronomical parameters.  Code adapted from ETERNA.
}
\references{
Wenzel, H.-G. (1996): The nanogal software: Earth tide data
processing package ETERNA 3.3. Bulletin d'Informations
Marees Terrestres vol. 124, 9425-9439, Bruxelles 1996.
}
\author{
Jonathan Kennel, \email{jkennel@uoguelph.ca}
}
