% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{astro_der}
\alias{astro_der}
\title{astro_der}
\usage{
astro_der(t_astro, simon)
}
\arguments{
\item{t_astro}{astronomical time}

\item{simon}{coefficient matrix}
}
\value{
derivative astonomical parameters
}
\description{
Calculate derivatives of astronomical parameters.  Code adapted from ETERNA.
}
\references{
Wenzel, H.-G. (1996): The nanogal software: Earth tide data
processing package ETERNA 3.3. Bulletin d'Informations
Marees Terrestres vol. 124, 9425-9439, Bruxelles 1996.
}
\author{
Jonathan Kennel, \email{jkennel@uoguelph.ca}
}
