% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{ddt}
\format{A \code{data.frame} The columns are:
\describe{
 \item{\code{year}}{year}
 \item{\code{month}}{month}
 \item{\code{day}}{day}
 \item{\code{ddt}}{ddt}
}}
\usage{
ddt
}
\description{
ddt
}
\examples{
\dontrun{
utils::data(ddt)
}
}
\keyword{datasets}
