% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_t.R
\name{delta_t}
\alias{delta_t}
\title{delta_t}
\usage{
delta_t(utc)
}
\arguments{
\item{utc}{datetime in utc}
}
\value{
seconds difference  32.184s + (TAI - UTC) - (UT1 - UTC)
}
\description{
delta_t returns the delta t
data from http://maia.usno.navy.mil/
}
