% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dut1}
\alias{dut1}
\title{dut1}
\format{A \code{data.frame} The columns are:
\describe{
 \item{\code{datetime}}{UTC time}
 \item{\code{ddt}}{TDT-UTC}
 \item{\code{ut1_utc}}{UT1-UTC}
 \item{\code{tai_utc}}{TAI - UTC}
 \item{\code{lod}}{length of day correction}
}}
\usage{
dut1
}
\description{
dut1
}
\examples{
\donttest{
utils::data(dut1)
}
}
\references{
http://hpiers.obspm.fr/eop-pc/index.php
}
\keyword{datasets}
