% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eterna_wavegroups}
\alias{eterna_wavegroups}
\title{Hartmann and Wenzel (1995) (ETERNA 3.4) wavegroups}
\format{A \code{data.frame} The columns are:
\describe{
 \item{\code{name}}{wave group name}
 \item{\code{start}}{lower frequency of the wave group}
 \item{\code{end}}{higher frequency of the wave group}
 \item{\code{time}}{applicable to data of what length}
}}
\usage{
eterna_wavegroups
}
\description{
Hartmann and Wenzel (1995) (ETERNA 3.4) wavegroups
}
\examples{
\dontrun{
utils::data(eterna_wavegroups)
summary(eterna_wavegroups)
}
}
\keyword{datasets}
