% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dut1}
\alias{get_dut1}
\title{get_dut1}
\usage{
get_dut1()
}
\value{
information on time conversions and leap seconds
}
\description{
Downloads data from http://maia.usno.navy.mil/ and updates package data
58.3092 from Monthly determinations of Delta T -- values of (TT - UT1) since 1973
}
