% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_iers}
\alias{get_iers}
\title{get_iers}
\usage{
get_iers()
}
\value{
\code{data.frame} of earth orientation parameters with the following 
columns: datetime, ddt, ut1_utc, lod, x, y, dx, dy.
}
\description{
\code{get_iers} returns a \code{data.frame} of earth orientation 
parameters from (1962-present).  This function requires an active internet connection. 
Bulletins A and B are combined giving precedence to B. 
Approximately (~ 7 MB) of data are downloaded.
}
\examples{
\dontrun{
eop <- get_iers()
}

}
