% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_main_frequency.R
\name{get_main_frequency}
\alias{get_main_frequency}
\title{get_main_frequency}
\usage{
get_main_frequency(start, end)
}
\arguments{
\item{start}{the starting frequency in cycles per day (numeric)}

\item{end}{the ending frequency in cycles per day (numeric)}
}
\value{
the main frequency between start and end
}
\description{
Get the frequency of the wave with the maximum amplitude in a range.
}
