% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gravity_station.R
\name{gravity_station}
\alias{gravity_station}
\title{gravity_station}
\usage{
gravity_station(latitude, elevation)
}
\arguments{
\item{latitude}{the station latitude in radians}

\item{elevation}{the station elevation in meters}
}
\value{
gravity for location in m/s^2
}
\description{
\code{gravity_station} returns the estimated gravity for a latitude.
http://the-mostly.ru/misc/local_gravity_online_calculator.html
}
