% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{legendre}
\alias{legendre}
\title{legendre}
\usage{
legendre(l_max, x)
}
\arguments{
\item{l_max}{maximum degree}

\item{x}{the x value}
}
\value{
legendre polynomials
}
\description{
Calculate legendre
}
