% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{legendre_bh}
\alias{legendre_bh}
\title{legendre_bh}
\usage{
legendre_bh(l, m, x, csphase = -1L)
}
\arguments{
\item{l}{maximum degree}

\item{m}{the m value}

\item{x}{the x value}

\item{csphase}{the csphase value}
}
\value{
legendre scale
}
\description{
legendre function
}
