% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{legendre_deriv_bh}
\alias{legendre_deriv_bh}
\title{legendre_deriv_bh}
\usage{
legendre_deriv_bh(l, m, x)
}
\arguments{
\item{l}{maximum degree}

\item{m}{the x value}

\item{x}{the x value}
}
\value{
legendre scale
}
\description{
Scaling for legendre function
}
