% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/love.R
\name{love}
\alias{love}
\title{love}
\usage{
love(latitude, elevation)
}
\arguments{
\item{latitude}{latitude the station latitude}

\item{elevation}{elevation the station elevation}
}
\value{
list of love numbers
}
\description{
returns the love numbers - direct port of eterna 3.40 code
http://igets.u-strasbg.fr/soft_and_tool.php
}
