% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{scale_legendre_bh}
\alias{scale_legendre_bh}
\title{scale_legendre_bh}
\usage{
scale_legendre_bh(l, m)
}
\arguments{
\item{l}{maximum degree}

\item{m}{the x value}
}
\value{
legendre scale
}
\description{
Scaling for legendre function
}
