% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simon_coef_1994}
\alias{simon_coef_1994}
\title{Simon 1994 astronomical constants}
\format{A \code{data.frame} The columns are:
\describe{
 \item{\code{j}}{7 = mercury, 8 = venus, 9 = mars, 10 = Jupiter, 11 = Saturn}
 \item{\code{constant}}{constant}
 \item{\code{t1}}{t1}
 \item{\code{t2}}{t2}
 \item{\code{t3}}{t3}
 \item{\code{t4}}{t4}
}}
\usage{
simon_coef_1994
}
\description{
Simon 1994 astronomical constants
}
\examples{
\donttest{
utils::data(simon_coef_1994)
summary(simon_coef_1994)
}
}
\references{
Simon JL, Bretagnon P, Chapront J, Chapront-Touzè M, Francou G, Laskar J (1994) Numerical expressions for precession for- mulae and mean elements for the Moon and planets. Astron Astrophys 282:663–683
}
\keyword{datasets}
