% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{time_der_mat}
\alias{time_der_mat}
\title{time_der_mat}
\usage{
time_der_mat(time)
}
\arguments{
\item{time}{the times to calculate the derivatives of the astrological parameters}
}
\value{
matrix of time derivatives
}
\description{
Matrix to multiply by astrological.  Code adapted from ETERNA.
}
\references{
Wenzel, H.-G. (1996): The nanogal software: Earth tide data
processing package ETERNA 3.3. Bulletin d'Informations
Marees Terrestres vol. 124, 9425-9439, Bruxelles 1996.
}
\author{
Jonathan Kennel, \email{jkennel@uoguelph.ca}
}
