% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ut1_utc.R
\name{ut1_utc}
\alias{ut1_utc}
\title{ut1_utc}
\usage{
ut1_utc(utc)
}
\arguments{
\item{utc}{datetime in utc}
}
\value{
seconds difference UT1-UTC
}
\description{
\code{tai_ut1} gets seconds difference UT1-UTC (valid after 1962)
}
