% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_date.R
\name{utc_julian}
\alias{utc_julian}
\title{utc_julian}
\usage{
utc_julian(utc)
}
\arguments{
\item{utc}{datetime in utc}
}
\value{
julian datetime
}
\description{
\code{utc_julian} returns the julian datetime from UTC
}
