% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_date.R
\name{utc_julian_2000}
\alias{utc_julian_2000}
\title{utc_julian_2000}
\usage{
utc_julian_2000(utc)
}
\arguments{
\item{utc}{datetime in utc}
}
\value{
julian datetime
}
\description{
\code{utc_julian_2000} returns the j2000 datetime from UTC
}
