% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthtide-package.R
\docType{package}
\name{earthtide-package}
\alias{earthtide-package}
\alias{_PACKAGE}
\title{earthtide: R port of the earth tide processing package ETERNA 
(by Hans-Georg Wenzel) including the Kudryavtsev wave catalog.}
\description{
The goal of this package is to generate synthetic earth tides for use in the
R programming language and in particular environmental models. Code was 
parallized and refactored to minimize duplication, and to allow for future 
improvements
}
\details{
You can learn about the earthtide package in the vignettes:
`browseVignettes(package = "earthtide")`
}
\author{
\strong{Maintainer}: Jonathan Kennel \email{jkennel@uoguelph.ca} [translator]

Authors:
\itemize{
  \item Wenzel Hans-Georg
}

Other contributors:
\itemize{
  \item Beth Parker \email{bparker@uoguelph.ca} [thesis advisor]
}

}
