% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{eterna_wavegroups}
\alias{eterna_wavegroups}
\title{Hartmann and Wenzel (1995) (ETERNA 3.4) wavegroups}
\format{A \code{data.frame} The columns are:
\describe{
 \item{\code{name}}{wave group name}
 \item{\code{start}}{lowest frequency of the wave group}
 \item{\code{end}}{highest frequency of the wave group}
 \item{\code{time}}{applicable to data of what length}
}}
\usage{
eterna_wavegroups
}
\description{
This data.frame contains wavegroups for different data time
spans.  The wavegroups should be subset prior to use and the 'time' column
provides guidelines based on your input time span.
}
\examples{
utils::data(eterna_wavegroups)
}
\keyword{datasets}
