% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dut1}
\alias{get_dut1}
\title{get_dut1}
\usage{
get_dut1()
}
\value{
information on time conversions and leap seconds
}
\description{
Downloads earth orientation data from http://hpiers.obspm.fr/eop-pc/index.php
}
